Refactored our code into more reusable modules
