+++ tasks/toArray.js
@@ -1,0 +1,7 @@
+var types = require("dis-isa");
+
+function toArray(data) {
+  return types.isArray(data) ? data : [data];
+}
+
+module.exports = toArray;
